chore: move filter chips hook out
