+++ src/swap/SwapScreen.tsx
@@ -18,14 +18,13 @@ import BackButton from 'src/components/BackButton'
 import BottomSheet, { BottomSheetRefType } from 'src/components/BottomSheet'
 import BottomSheetInLineNotification from 'src/components/BottomSheetInLineNotification'
 import Button, { BtnSizes, BtnTypes } from 'src/components/Button'
-import { FilterChip } from 'src/components/FilterChipsCarousel'
 import InLineNotification, { Severity } from 'src/components/InLineNotification'
 import TokenBottomSheet, {
   TokenBalanceItemOption,
   TokenPickerOrigin,
 } from 'src/components/TokenBottomSheet'
 import CustomHeader from 'src/components/header/CustomHeader'
-import { SWAP_LEARN_MORE, TOKEN_MIN_AMOUNT } from 'src/config'
+import { SWAP_LEARN_MORE } from 'src/config'
 import { FiatExchangeFlow } from 'src/fiatExchanges/utils'
 import { getLocalCurrencyCode } from 'src/localCurrency/selectors'
 import { navigate } from 'src/navigator/NavigationService'
@@ -33,9 +32,9 @@ import { Screens } from 'src/navigator/Screens'
 import { StackParamList } from 'src/navigator/types'
 import useSelector from 'src/redux/useSelector'
 import { NETWORK_NAMES } from 'src/shared/conts'
-import { getDynamicConfigParams, getExperimentParams, getFeatureGate } from 'src/statsig'
+import { getDynamicConfigParams, getExperimentParams } from 'src/statsig'
 import { DynamicConfigs, ExperimentConfigs } from 'src/statsig/constants'
-import { StatsigDynamicConfigs, StatsigExperiments, StatsigFeatureGates } from 'src/statsig/types'
+import { StatsigDynamicConfigs, StatsigExperiments } from 'src/statsig/types'
 import colors from 'src/styles/colors'
 import fontStyles, { typeScale } from 'src/styles/fonts'
 import { Spacing } from 'src/styles/styles'
@@ -44,23 +43,20 @@ import PreparedTransactionsReviewBottomSheet from 'src/swap/PreparedTransactions
 import SwapAmountInput from 'src/swap/SwapAmountInput'
 import SwapTransactionDetails from 'src/swap/SwapTransactionDetails'
 import { getSwapTxsAnalyticsProperties } from 'src/swap/getSwapTxsAnalyticsProperties'
-import {
-  currentSwapSelector,
-  lastSwappedSelector,
-  priceImpactWarningThresholdSelector,
-} from 'src/swap/selectors'
+import { currentSwapSelector, priceImpactWarningThresholdSelector } from 'src/swap/selectors'
 import { swapStart } from 'src/swap/slice'
 import { Field, SwapAmount } from 'src/swap/types'
+import useFilterChips from 'src/swap/useFilterChips'
 import useSwapQuote, { QuoteResult } from 'src/swap/useSwapQuote'
 import { useSwappableTokens, useTokenInfo, useTokensWithTokenBalance } from 'src/tokens/hooks'
 import { feeCurrenciesWithPositiveBalancesSelector, tokensByIdSelector } from 'src/tokens/selectors'
 import { TokenBalance } from 'src/tokens/slice'
 import { getSupportedNetworkIdsForSwap, getTokenId } from 'src/tokens/utils'
-import { Network, NetworkId } from 'src/transactions/types'
+import { NetworkId } from 'src/transactions/types'
 import Logger from 'src/utils/Logger'
 import { getFeeCurrencyAndAmounts } from 'src/viem/prepareTransactions'
 import { getSerializablePreparedTransactions } from 'src/viem/preparedTransactionSerialization'
-import networkConfig, { networkIdToNetwork } from 'src/web3/networkConfig'
+import networkConfig from 'src/web3/networkConfig'
 import { v4 as uuidv4 } from 'uuid'
 
 const TAG = 'SwapScreen'
@@ -215,68 +211,6 @@ function getNetworkFee(quote: QuoteResult | null, networkId?: NetworkId) {
 }
 
 type Props = NativeStackScreenProps<StackParamList, Screens.SwapScreenWithBack>
-
-const useFilterChips = (selectingField: Field | null) => {
-  const { t } = useTranslation()
-  const showSwapTokenFilters = getFeatureGate(StatsigFeatureGates.SHOW_SWAP_TOKEN_FILTERS)
-  const recentlySwappedTokens = useSelector(lastSwappedSelector)
-  const popularTokens: string[] = [] // TODO
-  const supportedNetworkIds = getSupportedNetworkIdsForSwap()
-
-  const preSelectedFilterChipIds =
-    selectingField === Field.FROM ? ['my-tokens'] : ['my-tokens', 'popular']
-
-  const filterChips: FilterChip<TokenBalance>[] = useMemo(() => {
-    if (!showSwapTokenFilters) {
-      return []
-    }
-
-    return [
-      {
-        id: 'my-tokens',
-        name: t('tokenBottomSheet.filters.myTokens'),
-        filterFn: (token: TokenBalance) => token.balance.gte(TOKEN_MIN_AMOUNT),
-      },
-      {
-        id: 'popular',
-        name: t('tokenBottomSheet.filters.popular'),
-        filterFn: (token: TokenBalance) => popularTokens.includes(token.tokenId),
-      },
-      {
-        id: 'recently-swapped',
-        name: t('tokenBottomSheet.filters.recentlySwapped'),
-        filterFn: (token: TokenBalance) => recentlySwappedTokens.includes(token.tokenId),
-      },
-      ...(supportedNetworkIds.length > 1
-        ? [
-            {
-              id: 'celo-network',
-              name: t('tokenBottomSheet.filters.celo'),
-              filterFn: (token: TokenBalance) =>
-                networkIdToNetwork[token.networkId] === Network.Celo,
-            },
-          ]
-        : []),
-      ...(supportedNetworkIds.includes(NetworkId['ethereum-mainnet'])
-        ? [
-            {
-              id: 'ethereum-network',
-              name: t('tokenBottomSheet.filters.ethereum'),
-              filterFn: (token: TokenBalance) =>
-                networkIdToNetwork[token.networkId] === Network.Ethereum,
-            },
-          ]
-        : []),
-    ]
-  }, [t, recentlySwappedTokens, popularTokens, supportedNetworkIds, showSwapTokenFilters])
-
-  return {
-    filterChips,
-    preSelectedFilterChips: filterChips.filter((chip) =>
-      preSelectedFilterChipIds.includes(chip.id)
-    ),
-  }
-}
 
 export function SwapScreen({ route }: Props) {
   const { t } = useTranslation()
